bac-analyzer-api/api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pyyaml==6.0.1